Moved some folder inside Application
`Application/Repository` and `Application/Exception` has been
moved inside `Application/Group` folder.
