+++ config/packages/prooph_event_store.yaml
@@ -14,4 +14,4 @@ services:
         public: false
 
     Prooph\EventSourcing\EventStoreIntegration\AggregateTranslator: ~
-    App\Application\Repository\Groups: '@App\Infrastructure\Repository\EventStoreGroups'
+    App\Application\Group\Repository\Groups: '@App\Infrastructure\Repository\EventStoreGroups'
