+++ src/Application/Group/Command/AddGroupHandler.php
@@ -11,7 +11,7 @@
 
 namespace App\Application\Group\Command;
 
-use App\Application\Repository\Groups;
+use App\Application\Group\Repository\Groups;
 use App\Domain\Group\Model\Group;
 
 class AddGroupHandler
